unified configuration API (_globals/_locals -> config)
Previously it was really unclear as to where a configuration setting was
actually coming from. Is it a global? Is it a local? Sometimes a setting
in globals would manually override itself with the value it pulled out
of locals, so even if you overrode it in locals you'd still need to
*import* it from globals. Finally, these two modules' names both
collided with Python builtins and had to be aliased every time we
imported them.

Now there's just config. You import it. You use its attributes. If you
put something in your locals it overrides the thing in globals.